refactor(format): split fun renderer owners
